Fix QA P1 checkpoint findings: memory footprint <= C zlib + accurate CI/docs/deck
Resolves 6 QA findings (4 Major, 2 Minor), all tracing to the AAP memory-footprint gate:

P1 (Major, Performance): Reduce engine state to <= C zlib. Box gzhead/window/pending_buf and narrow 15 non-position scalar fields (usize->u32) in DeflateState (6136->5952 = C parity; engine 268096); box head + window in InflateState (1488->1376; engine 7152 <= C 7160). opt_len/static_len kept usize and all LZ77 position fields kept usize/isize to preserve 64-bit wrapping and byte-identical output. Fixes no-std Vec import fallout. Adds tests/memory_footprint.rs (counting-allocator runtime gate) + in-crate size_of gate tests for both structs.

C1 (Major, CI): Add a memory-footprint enforcement step to the ci.yml perf job and rewrite the header to enumerate all four AAP 0.3.3 gates (3 throughput via Criterion + memory via the allocation test).

D1 (Major, Docs): README benchmarks section now reports the measured memory result (deflate parity 268096 B; inflate 7152 B, -8 B) and accurately describes CI enforcing all four gates.

D2 (Minor, Docs): Refresh doc/project-guide.md to current state - branch, test counts (500 = 296 unit + 139 integration + 65 doc, 5 ignored doctests), no-default-features exits 0, byte-identity/performance validated, risk/task tables updated.

F1 (Major, Frontend): Executive presentation slide 11 memory KPI and caption now reflect the verified passing state (measured at parity; all four gates enforced in CI).

F2 (Minor, A11y): Enlarge deck nav controls to 44x44 (WCAG 2.5.5), add a visible token-based :focus-visible indicator, and add an accessible portrait hint documenting landscape/desktop viewing for the fixed 1920x1080 canvas.

Verification: full suite 500 passing, clippy -D warnings + fmt clean, all four perf gates pass (CRC 55x, DEFLATE 0.86x, INFLATE 1.01x, memory <= C zlib), FFI C drop-in round-trip OK, deck renders clean. Byte-identical output preserved.

+++ src/deflate/fast.rs
@@ -109,7 +109,9 @@ pub(crate) fn deflate_fast(s: &mut DeflateStream, flush: FlushMode) -> BlockStat
             // Insert new strings in the hash table only if the match length is
             // not too large. This saves time but degrades compression.
             // (`max_lazy_match` doubles as C's `max_insert_length`.)
-            if s.state.match_length <= s.state.max_lazy_match && s.state.lookahead >= MIN_MATCH {
+            if s.state.match_length <= s.state.max_lazy_match as usize
+                && s.state.lookahead >= MIN_MATCH
+            {
                 s.state.match_length -= 1; // string at strstart already in table
                 // C `do { strstart++; INSERT_STRING(...); } while(--match_length)`:
                 // decrement happens before the zero test, so this runs the body
